Initial handling of facet constraint
Only works if 1) it is a range constraint, and 2) the name matches a
json-property.

Obviously, we need to do much better than this, and we will. :-)

Also, start to use queryBuilder

Thinking of transitioning completely away from stored query options,
which seems to be the same direction that Engineering is taking, and why
they are not very supported in the Node API interface