User login works for search route
Note that I removed guestAccess and defaultUser/defaultPassword, at
least for now. This was simplifying, and we should discuss whether and
how to support that use case. It may be a different auth strategy
altogether. I am afraid that leaving it as a backup is a potential
security concern. A malevolent user could reach a route on the
middle tier, and thus gain access to the privileges of the default user.