Create some initial documentation around default-routes
Includes in-line comments in the user-owned routes file, and a README in
the default-routes library.

Still need to add top-level documentation to muir-node to emphasize that
the routes file is the place to start.

Also need to document defaultCRUDRoute

It is clear that we are still missing big pieces of configuration for
the routes, including the ability to override parsing, executing, etc,
for search.

GROVE-345