GROVE-489: call constraint functions in consistent way
We were binding the `this` of geospatialConstraint() because it in turn
calls this.geospatialValues(), which is an exported function so it is
accessible to some tests. The other constraint functions don't really
need it, but this commit makes it so that we call them in a consistent
way, to avoid confusion later.